Development

See merge request sanderstad/PSDatabaseClone!6

+++ functions/Remove-PSDCImage.ps1
@@ -90,22 +90,28 @@
 
     begin {
 
+        # Get the module configurations
+        $pdcSqlInstance = Get-PSFConfigValue -FullName psdatabaseclone.database.Server
+        $pdcDatabase = Get-PSFConfigValue -FullName psdatabaseclone.database.name
+        if (-not $pdcCredential) {
+            $pdcCredential = Get-PSFConfigValue -FullName psdatabaseclone.database.credential -Fallback $null
+        }
+        else {
+            $pdcCredential = $PSDCSqlCredential
+        }
+
         # Test the module database setup
         try {
-            Test-PSDCConfiguration -SqlCredential $PSDCSqlCredential -EnableException
+            Test-PSDCConfiguration -SqlCredential $pdcCredential -EnableException
         }
         catch {
             Stop-PSFFunction -Message "Something is wrong in the module configuration" -ErrorRecord $_ -Continue
         }
 
-        # Get the database values
-        $pdcSqlInstance = Get-PSFConfigValue -FullName psdatabaseclone.database.server
-        $pdcDatabase = Get-PSFConfigValue -FullName psdatabaseclone.database.name
-
         Write-PSFMessage -Message "Started removing database images" -Level Verbose
 
         # Get all the items
-        $items = Get-PSDCImage -PSDCSqlCredential $PSDCSqlCredential
+        $items = Get-PSDCImage
 
         if ($ImageID) {
             Write-PSFMessage -Message "Filtering image ids" -Level Verbose
@@ -148,43 +154,57 @@
             # Loop through each of the results
             foreach ($item in $image.Group) {
 
+                # Make up the data from the network path
+                try {
+                    [uri]$uri = New-Object System.Uri($item.ImageLocation)
+                    $uriHost = $uri.Host
+                }
+                catch {
+                    Stop-PSFFunction -Message "The image location $ImageNetworkPath is not valid" -ErrorRecord $_ -Target $ImageNetworkPath
+                    return
+                }
+
+                # Setup the computer object
+                $computer = [PsfComputer]$uriHost
+
                 $query = "
-                SELECT c.CloneLocation,
-                        c.AccessPath,
-                        c.SqlInstance,
-                        c.DatabaseName,
-                        h.HostName,
-                        h.IPAddress,
-                        h.FQDN,
-                        i.ImageLocation
-                FROM dbo.Image as i
-                    INNER JOIN dbo.Clone AS c
-                        ON c.ImageID = i.ImageID
-                    INNER JOIN dbo.Host AS h
-                        ON h.HostID = c.HostID
-                WHERE i.ImageID = $($item.ImageID)
-                ORDER BY h.HostName;
-            "
+                    SELECT c.CloneLocation,
+                            c.AccessPath,
+                            c.SqlInstance,
+                            c.DatabaseName,
+                            h.HostName,
+                            h.IPAddress,
+                            h.FQDN,
+                            i.ImageLocation
+                    FROM dbo.Image as i
+                        INNER JOIN dbo.Clone AS c
+                            ON c.ImageID = i.ImageID
+                        INNER JOIN dbo.Host AS h
+                            ON h.HostID = c.HostID
+                    WHERE i.ImageID = $($item.ImageID)
+                    ORDER BY h.HostName;
+                "
 
                 # Try to get the neccesary info from the EasyClone database
                 try {
                     Write-PSFMessage -Message "Retrieving data for image '$($item.Name)'" -Level Verbose
                     $results = @()
-                    $results += Invoke-DbaSqlQuery -SqlInstance $pdcSqlInstance -SqlCredential $PSDCSqlCredential -Database $pdcDatabase -Query $query
+                    $results += Invoke-DbaSqlQuery -SqlInstance $pdcSqlInstance -SqlCredential $pdcCredential -Database $pdcDatabase -Query $query
 
                     # Check the results
                     if ($results.Count -ge 1) {
 
                         # Loop through the results
                         foreach ($result in $results) {
-
-                            # Remove the clones for the host
-                            try {
-                                Write-PSFMessage -Message "Removing clones for host $($result.HostName) and database $($result.DatabaseName)" -Level Verbose
-                                Remove-PSDCClone -HostName $result.HostName -Database $result.DatabaseName -PSDCSqlCredential $PSDCSqlCredential -Credential $Credential -Confirm:$false
-                            }
-                            catch {
-                                Stop-PSFFunction -Message "Couldn't remove clones from host $($result.HostName)" -ErrorRecord $_ -Target $result -Continue
+                            if ($PSCmdlet.ShouldProcess($item.CloneID, "Removing clone $($result.CloneLocation) from $($result.HostName)")) {
+                                # Remove the clones for the host
+                                try {
+                                    Write-PSFMessage -Message "Removing clones for host $($result.HostName) and database $($result.DatabaseName)" -Level Verbose
+                                    Remove-PSDCClone -HostName $result.HostName -Database $result.DatabaseName -PSDCSqlCredential $pdcCredential -Credential $Credential -Confirm:$false
+                                }
+                                catch {
+                                    Stop-PSFFunction -Message "Couldn't remove clones from host $($result.HostName)" -ErrorRecord $_ -Target $result -Continue
+                                }
                             }
                         }
                     }
@@ -196,28 +216,47 @@
                     Stop-PSFFunction -Message "Couldn't retrieve clone records for host $($result.HostName)" -ErrorRecord $_  -Target $hst -Continue
                 }
 
-                # Remove the image from the file system
-                try {
-                    if (Test-Path -Path $item.ImageLocation -Credential $Credential) {
-                        Write-PSFMessage -Message "Removing image '$($item.ImageLocation)' from file system" -Level Verbose
-                        $null = Remove-Item -Path $item.ImageLocation -Credential $Credential -Force:$Force
+                if ($PSCmdlet.ShouldProcess($item.ImageLocation, "Removing image from system")) {
+                    # Remove the image from the file system
+                    try {
+                        if ($computer.IsLocalhost) {
+                            if (Test-Path -Path $item.ImageLocation -Credential $Credential) {
+                                Write-PSFMessage -Message "Removing image '$($item.ImageLocation)' from file system" -Level Verbose
+
+                                $null = Remove-Item -Path $item.ImageLocation -Credential $Credential -Force:$Force
+                            }
+                            else {
+                                Write-PSFMessage -Message "Couldn't find image $($item.ImageLocation)" -Level Verbose
+                            }
+                        }
+                        else {
+                            $command = [scriptblock]::Create("Test-Path -Path $($item.ImageLocation)")
+                            $result = Invoke-PSFCommand -ComputerName $computer -ScriptBlock $command -Credential $Credential
+
+                            if ($result) {
+                                $command = [scriptblock]::Create("Remove-Item -Path $($item.ImageLocation) -Force")
+                                $result = Invoke-PSFCommand -ComputerName $computer -ScriptBlock $command -Credential $Credential
+                            }
+                            else {
+                                Write-PSFMessage -Message "Couldn't find image $($item.ImageLocation)" -Level Verbose
+                            }
+                        }
                     }
-                    else {
-                        Write-PSFMessage -Message "Couldn't find image $($item.ImageLocation)" -Level Verbose
+                    catch {
+                        Stop-PSFFunction -Message "Couldn't remove image '$($item.ImageLocation)' from file system" -ErrorRecord $_ -Target $result
                     }
-                }
-                catch {
-                    Stop-PSFFunction -Message "Couldn't remove image '$($item.ImageLocation)' from file system" -ErrorRecord $_ -Target $result
                 }
 
-                # Remove the image from the database
-                try {
-                    $query = "DELETE FROM dbo.Image WHERE ImageID = $($item.ImageID)"
+                if ($PSCmdlet.ShouldProcess($item.ImageLocation, "Removing image from database")) {
+                    # Remove the image from the database
+                    try {
+                        $query = "DELETE FROM dbo.Image WHERE ImageID = $($item.ImageID)"
 
-                    $null = Invoke-DbaSqlQuery -SqlInstance $pdcSqlInstance -SqlCredential $PSDCSqlCredential -Database $pdcDatabase -Query $query
-                }
-                catch {
-                    Stop-PSFFunction -Message "Couldn't remove image '$($item.ImageLocation)' from database" -ErrorRecord $_ -Target $query
+                        $null = Invoke-DbaSqlQuery -SqlInstance $pdcSqlInstance -SqlCredential $pdcCredential -Database $pdcDatabase -Query $query
+                    }
+                    catch {
+                        Stop-PSFFunction -Message "Couldn't remove image '$($item.ImageLocation)' from database" -ErrorRecord $_ -Target $query
+                    }
                 }
 
             } # End for each item in group
